Refactor, clean, and organize code
