+++ ipld/hamt/src/bitfield.rs
@@ -1,3 +1,6 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
 use byteorder::{BigEndian, ByteOrder};
 use forest_encoding::{
     de::{Deserialize, Deserializer},
