+++ ipld/hamt/src/hamt.rs
@@ -1,18 +1,14 @@
-use crate::bitfield::Bitfield;
-use crate::hash::Hash;
-use crate::Error;
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
+use crate::node::Node;
+use crate::{Error, Hash};
 use cid::Cid;
-use forest_encoding::{de::Deserializer, ser::Serializer};
 use ipld_blockstore::BlockStore;
-use lazycell::AtomicLazyCell;
-use murmur3::murmur3_x64_128::MurmurHasher;
-use replace_with::replace_with;
 use serde::de::DeserializeOwned;
-use serde::{Deserialize, Serialize};
+use serde::Serialize;
 use std::borrow::Borrow;
 use std::ops::Index;
-
-const MAX_ARRAY_WIDTH: usize = 3;
 
 /// Implementation of the HAMT data structure for IPLD.
 #[derive(Debug)]
@@ -24,93 +20,6 @@ pub struct Hamt<'a, K, V, S: BlockStore> {
 impl<'a, K: PartialEq, V: PartialEq, S: BlockStore> PartialEq for Hamt<'a, K, V, S> {
     fn eq(&self, other: &Self) -> bool {
         self.root == other.root
-    }
-}
-
-#[derive(Debug, Clone, PartialEq, Eq)]
-struct Node<K, V> {
-    bitfield: Bitfield,
-    pointers: Vec<Pointer<K, V>>,
-}
-
-impl<K, V> Serialize for Node<K, V>
-where
-    K: Serialize,
-    V: Serialize,
-{
-    fn serialize<S>(&self, serializer: S) -> Result<<S as Serializer>::Ok, <S as Serializer>::Error>
-    where
-        S: Serializer,
-    {
-        (&self.bitfield, &self.pointers).serialize(serializer)
-    }
-}
-
-impl<'de, K, V> Deserialize<'de> for Node<K, V>
-where
-    K: DeserializeOwned,
-    V: DeserializeOwned,
-{
-    fn deserialize<D>(deserializer: D) -> Result<Self, D::Error>
-    where
-        D: Deserializer<'de>,
-    {
-        let (bitfield, pointers) = Deserialize::deserialize(deserializer)?;
-        Ok(Node { bitfield, pointers })
-    }
-}
-
-type HashedKey = [u8; 16];
-
-// TODO: make Pointer an enum once things are working
-#[derive(Debug, Clone, Serialize, Deserialize)]
-#[serde(bound(deserialize = "KeyValuePair<K, V>: DeserializeOwned"))]
-struct Pointer<K, V> {
-    #[serde(rename = "v", skip_serializing_if = "Vec::is_empty")]
-    kvs: Vec<KeyValuePair<K, V>>,
-    #[serde(rename = "l", skip_serializing_if = "Option::is_none")]
-    link: Option<Cid>,
-    #[serde(skip)]
-    cache: AtomicLazyCell<Node<K, V>>,
-}
-
-impl<K: PartialEq, V: PartialEq> PartialEq for Pointer<K, V> {
-    fn eq(&self, other: &Self) -> bool {
-        self.kvs == other.kvs && self.link == other.link
-    }
-}
-
-impl<K: Eq, V: Eq> Eq for Pointer<K, V> {}
-
-impl<K, V> Default for Pointer<K, V> {
-    fn default() -> Self {
-        Pointer {
-            kvs: Vec::new(),
-            link: None,
-            cache: AtomicLazyCell::new(),
-        }
-    }
-}
-
-#[derive(Debug, Default, Clone, Serialize, Deserialize, PartialEq, Eq)]
-struct KeyValuePair<K, V>(K, V);
-
-impl<K, V> KeyValuePair<K, V> {
-    pub fn key(&self) -> &K {
-        &self.0
-    }
-
-    pub fn value(&self) -> &V {
-        &self.1
-    }
-}
-
-impl<K, V> Default for Node<K, V> {
-    fn default() -> Self {
-        Node {
-            bitfield: Bitfield::zero(),
-            pointers: Vec::new(),
-        }
     }
 }
 
@@ -131,7 +40,7 @@ where
     pub fn from_link(cid: &Cid, store: &'a S) -> Result<Self, Error> {
         match store.get(cid)? {
             Some(root) => Ok(Hamt { root, store }),
-            None => Err(Error::Custom("No node found".to_owned())),
+            None => Err(Error::Custom("No node found")),
         }
     }
 
@@ -245,384 +154,6 @@ where
 
     pub fn is_empty(&self) -> bool {
         self.root.is_empty()
-    }
-}
-
-// TODO: should all operations be wrapped into Result?
-
-impl<K, V> Node<K, V>
-where
-    K: Hash + Eq + std::cmp::PartialOrd + Serialize + DeserializeOwned,
-    V: Serialize + DeserializeOwned,
-{
-    pub fn insert<S: BlockStore>(&mut self, key: K, value: V, store: &S) -> Option<V> {
-        match self.modify_value(Self::hash(&key), 0, key, value, store) {
-            Ok(res) => res,
-            Err(_) => None,
-        }
-    }
-
-    #[inline]
-    pub fn get<Q: ?Sized, S: BlockStore>(&self, k: &Q, store: &S) -> Option<&V>
-    where
-        K: Borrow<Q>,
-        Q: Eq + Hash,
-    {
-        self.search(k, store).map(|kv| kv.value())
-    }
-
-    #[inline]
-    pub fn remove_entry<Q: ?Sized, S>(&mut self, k: &Q, store: &S) -> Option<(K, V)>
-    where
-        K: Borrow<Q>,
-        Q: Eq + Hash,
-        S: BlockStore,
-    {
-        match self.rm_value(Self::hash(k), 0, k, store) {
-            Ok(res) => res,
-            Err(_) => None,
-        }
-    }
-
-    pub fn is_empty(&self) -> bool {
-        self.pointers.is_empty()
-    }
-
-    /// Search for a key.
-    #[inline]
-    fn search<Q: ?Sized, S: BlockStore>(&self, q: &Q, store: &S) -> Option<&KeyValuePair<K, V>>
-    where
-        K: Borrow<Q>,
-        Q: Eq + Hash,
-    {
-        self.get_value(Self::hash(q), 0, q, store)
-    }
-
-    fn get_value<Q: ?Sized, S: BlockStore>(
-        &self,
-        hashed_key: HashedKey,
-        depth: usize,
-        key: &Q,
-        store: &S,
-    ) -> Option<&KeyValuePair<K, V>>
-    where
-        K: Borrow<Q>,
-        Q: Eq + Hash,
-    {
-        assert!(depth < hashed_key.len(), "max depth reached");
-
-        let idx = hashed_key[depth];
-        if !self.bitfield.test_bit(idx) {
-            return None;
-        }
-
-        let cindex = self.index_for_bit_pos(idx);
-        let child = self.get_child(cindex);
-        if child.is_shard() {
-            match child.load_child(store) {
-                Ok(chnd) => chnd.get_value(hashed_key, depth + 1, key, store),
-                Err(_) => None,
-            }
-        } else {
-            child.kvs.iter().find(|kv| key.eq(kv.key().borrow()))
-        }
-    }
-
-    /// The hash function used to hash keys.
-    fn hash<X: ?Sized>(key: &X) -> HashedKey
-    where
-        X: Hash,
-    {
-        let mut hasher = MurmurHasher::default();
-        key.hash(&mut hasher);
-        hasher.finalize().into()
-    }
-
-    /// Internal method to modify values.
-    fn modify_value<S: BlockStore>(
-        &mut self,
-        hashed_key: HashedKey,
-        depth: usize,
-        key: K,
-        value: V,
-        store: &S,
-    ) -> Result<Option<V>, Error> {
-        if depth >= hashed_key.len() {
-            return Err(Error::Custom("Maximum depth reached".to_owned()));
-        }
-        let idx = hashed_key[depth];
-
-        // No existing values at this point.
-        if !self.bitfield.test_bit(idx) {
-            self.insert_child(idx, key, value);
-            return Ok(None);
-        }
-
-        let cindex = self.index_for_bit_pos(idx);
-        let child = self.get_child_mut(cindex);
-
-        if child.is_shard() {
-            let chnd = child.load_child_mut(store)?;
-            let v = chnd.modify_value(hashed_key, depth + 1, key, value, store)?;
-
-            return Ok(v);
-        }
-
-        // Update, if the key already exists.
-        if let Some(i) = child.kvs.iter().position(|p| p.key() == &key) {
-            let old_value = std::mem::replace(&mut child.kvs[i].1, value);
-            return Ok(Some(old_value));
-        }
-
-        // If the array is full, create a subshard and insert everything
-        if child.kvs.len() > MAX_ARRAY_WIDTH {
-            let mut sub = Node::default();
-            sub.modify_value(hashed_key, depth + 1, key, value, store)?;
-            let kvs = std::mem::replace(&mut child.kvs, Vec::new());
-            for p in kvs.into_iter() {
-                sub.modify_value(Self::hash(p.key()), depth + 1, p.0, p.1, store)?;
-            }
-
-            let link = store.put(&sub)?;
-            self.set_child(cindex, Pointer::from_link(link, sub));
-            return Ok(None);
-        }
-
-        // Otherwise insert the element into the array in order.
-        let max = child.kvs.len();
-        let idx = child
-            .kvs
-            .iter()
-            .position(|c| c.key() > &key)
-            .unwrap_or_else(|| max);
-
-        let np = KeyValuePair::new(key, value);
-        child.kvs.insert(idx, np);
-
-        Ok(None)
-    }
-
-    /// Internal method to delete entries.
-    pub fn rm_value<Q: ?Sized, S: BlockStore>(
-        &mut self,
-        hashed_key: HashedKey,
-        depth: usize,
-        key: &Q,
-        store: &S,
-    ) -> Result<Option<(K, V)>, Error>
-    where
-        K: Borrow<Q>,
-        Q: Hash + Eq,
-    {
-        if depth >= hashed_key.len() {
-            return Err(Error::Custom("Maximum depth reached".to_owned()));
-        }
-        let idx = hashed_key[depth];
-
-        // No existing values at this point.
-        if !self.bitfield.test_bit(idx) {
-            return Ok(None);
-        }
-
-        let cindex = self.index_for_bit_pos(idx);
-        let child = self.get_child_mut(cindex);
-
-        if child.is_shard() {
-            let chnd = child.load_child_mut(store)?;
-            let v = chnd.rm_value(hashed_key, depth + 1, key, store)?;
-
-            // CHAMP optimization, ensure trees look correct after deletion
-            return child.clean().map(|_| v);
-        }
-
-        // Delete value
-        for (i, p) in child.kvs.iter().enumerate() {
-            if key.eq(p.key().borrow()) {
-                let old = if child.kvs.len() == 1 {
-                    self.rm_child(cindex, idx).kvs.into_iter().nth(0).unwrap()
-                } else {
-                    child.kvs.remove(i)
-                };
-                return Ok(Some((old.0, old.1)));
-            }
-        }
-
-        Ok(None)
-    }
-
-    fn rm_child(&mut self, i: usize, idx: u8) -> Pointer<K, V> {
-        self.bitfield.clear_bit(idx);
-        self.pointers.remove(i)
-    }
-
-    fn insert_child(&mut self, idx: u8, key: K, value: V) {
-        let i = self.index_for_bit_pos(idx);
-        self.bitfield.set_bit(idx);
-        self.pointers
-            .insert(i as usize, Pointer::from_key_value(key, value))
-    }
-
-    fn set_child(&mut self, i: usize, pointer: Pointer<K, V>) {
-        self.pointers[i] = pointer;
-    }
-
-    fn index_for_bit_pos(&self, bp: u8) -> usize {
-        let mask = Bitfield::zero().set_bits_le(bp);
-        assert_eq!(mask.count_ones(), bp as usize);
-        let res = mask.and(&self.bitfield).count_ones();
-        res
-    }
-
-    fn get_child_mut(&mut self, i: usize) -> &mut Pointer<K, V> {
-        &mut self.pointers[i]
-    }
-
-    fn get_child<'a>(&'a self, i: usize) -> &'a Pointer<K, V> {
-        &self.pointers[i]
-    }
-}
-
-impl<K, V> KeyValuePair<K, V> {
-    pub fn new(key: K, value: V) -> Self {
-        KeyValuePair(key, value)
-    }
-}
-
-impl<K, V> Pointer<K, V>
-where
-    K: Serialize + DeserializeOwned,
-    V: Serialize + DeserializeOwned,
-{
-    pub fn from_link(link: Cid, node: Node<K, V>) -> Self {
-        let cache = AtomicLazyCell::new();
-        cache.fill(node).map_err(|_| ()).unwrap();
-
-        Pointer {
-            kvs: Vec::new(),
-            link: Some(link),
-            cache,
-        }
-    }
-
-    pub fn from_key_value(key: K, value: V) -> Self {
-        Pointer {
-            kvs: vec![KeyValuePair::new(key, value)],
-            link: None,
-            cache: AtomicLazyCell::new(),
-        }
-    }
-
-    pub fn from_kvpairs(kvs: Vec<KeyValuePair<K, V>>) -> Self {
-        Pointer {
-            kvs,
-            link: None,
-            cache: AtomicLazyCell::new(),
-        }
-    }
-
-    pub fn is_shard(&self) -> bool {
-        self.link.is_some()
-    }
-
-    fn load_child<S: BlockStore>(&self, store: &S) -> Result<&Node<K, V>, Error> {
-        if !self.cache.filled() {
-            if let Some(ref link) = self.link {
-                match store.get(link)? {
-                    Some(node) => {
-                        self.cache.fill(node).map_err(|_| ()).unwrap();
-                    }
-                    None => return Err(Error::Custom("node not found".to_owned())),
-                }
-            } else {
-                return Err(Error::Custom(
-                    "Cannot load child from non link node".to_owned(),
-                ));
-            }
-        }
-        Ok(self.cache.borrow().unwrap())
-    }
-
-    fn load_child_mut<S: BlockStore>(&mut self, store: &S) -> Result<&mut Node<K, V>, Error> {
-        if !self.cache.filled() {
-            if let Some(ref link) = self.link {
-                match store.get(link)? {
-                    Some(node) => {
-                        self.cache.fill(node).map_err(|_| ()).unwrap();
-                    }
-                    None => return Err(Error::Custom("node not found".to_owned())),
-                }
-            } else {
-                return Err(Error::Custom(
-                    "Cannot load child from non link node".to_owned(),
-                ));
-            }
-        }
-        Ok(self.cache.borrow_mut().unwrap())
-    }
-
-    /// Internal method to cleanup children, to ensure consisten tree representation
-    /// after deletes.
-    pub fn clean(&mut self) -> Result<(), Error> {
-        assert!(self.cache.filled());
-        let len = self.cache.borrow().unwrap().pointers.len();
-        if len <= 0 {
-            return Err(Error::Custom("Invalid HAMT".to_owned()));
-        }
-
-        replace_with(
-            self,
-            || panic!(),
-            |self_| {
-                match len {
-                    1 => {
-                        // TODO: investigate todo in go-hamt-ipld
-                        if self_.cache.borrow().unwrap().pointers[0].is_shard() {
-                            return self_;
-                        }
-
-                        self_
-                            .cache
-                            .into_inner()
-                            .unwrap()
-                            .pointers
-                            .into_iter()
-                            .nth(0)
-                            .unwrap()
-                    }
-                    1..=MAX_ARRAY_WIDTH => {
-                        let (total_lens, has_shards): (Vec<_>, Vec<_>) = self_
-                            .cache
-                            .borrow()
-                            .unwrap()
-                            .pointers
-                            .iter()
-                            .map(|p| (p.kvs.len(), p.is_shard()))
-                            .unzip();
-
-                        let total_len: usize = total_lens.iter().sum();
-                        let has_shards = has_shards.into_iter().any(|a| a);
-
-                        if total_len >= MAX_ARRAY_WIDTH || has_shards {
-                            return self_;
-                        }
-
-                        let chvals = self_
-                            .cache
-                            .into_inner()
-                            .unwrap()
-                            .pointers
-                            .into_iter()
-                            .map(|p| p.kvs)
-                            .flatten()
-                            .collect();
-
-                        Pointer::from_kvpairs(chvals)
-                    }
-                    _ => self_,
-                }
-            },
-        );
-        Ok(())
     }
 }
