+++ ipld/hamt/src/hash.rs
@@ -1,3 +1,6 @@
+// Copyright 2020 ChainSafe Systems
+// SPDX-License-Identifier: Apache-2.0, MIT
+
 use std::hash::Hasher;
 use std::{mem, slice};
